Remove unused parsing function and test
